# openapi.yaml
openapi: 3.0.1
info:
  title: BigBright ERP Domain API
  description: Core APIs for multi-company ERP platform
  license:
    name: Proprietary
  version: v1
...
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/v1/hr/employees/{id}:
    put:
      tags:
      - hr-controller
      operationId: updateEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseEmployeeDto'
    delete:
      tags:
      - hr-controller
      operationId: deleteEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
  /api/v1/factory/tasks/{id}:
    put:
      tags:
      - factory-controller
      operationId: updateTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FactoryTaskRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseFactoryTaskDto'
  /api/v1/factory/production-plans/{id}:
    put:
      tags:
      - factory-controller
      operationId: updatePlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionPlanRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductionPlanDto'
    delete:
      tags:
      - factory-controller
      operationId: deletePlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
...
  /api/v1/accounting/raw-materials/{id}:
    put:
      tags:
      - raw-material-controller
      operationId: updateRawMaterial
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawMaterialRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseRawMaterialDto'
    delete:
      tags:
      - raw-material-controller
      operationId: deleteRawMaterial
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
...
  /api/v1/raw-materials/intake:
    post:
      tags:
      - raw-material-controller
      operationId: intake
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawMaterialIntakeRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseRawMaterialBatchDto'
  /api/v1/raw-material-batches/{rawMaterialId}:
    get:
      tags:
      - raw-material-controller
      operationId: batches
      parameters:
      - name: rawMaterialId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListRawMaterialBatchDto'
    post:
      tags:
      - raw-material-controller
      operationId: createBatch
      parameters:
      - name: rawMaterialId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawMaterialBatchRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseRawMaterialBatchDto'
...
  /api/v1/hr/payroll-runs:
    get:
      tags:
      - hr-controller
      operationId: payrollRuns
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListPayrollRunDto'
    post:
      tags:
      - hr-controller
      operationId: createPayrollRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayrollRunRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePayrollRunDto'
  /api/v1/hr/leave-requests:
    get:
      tags:
      - hr-controller
      operationId: leaveRequests
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListLeaveRequestDto'
    post:
      tags:
      - hr-controller
      operationId: createLeaveRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequestRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLeaveRequestDto'
  /api/v1/hr/employees:
    get:
      tags:
      - hr-controller
      operationId: employees
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListEmployeeDto'
    post:
      tags:
      - hr-controller
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseEmployeeDto'
  /api/v1/factory/tasks:
    get:
      tags:
      - factory-controller
      operationId: tasks
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListFactoryTaskDto'
    post:
      tags:
      - factory-controller
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FactoryTaskRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseFactoryTaskDto'
  /api/v1/factory/production-plans:
    get:
      tags:
      - factory-controller
      operationId: plans
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListProductionPlanDto'
    post:
      tags:
      - factory-controller
      operationId: createPlan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionPlanRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductionPlanDto'
  /api/v1/factory/production-batches:
    get:
      tags:
      - factory-controller
      operationId: batches_1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListProductionBatchDto'
    post:
      tags:
      - factory-controller
      operationId: logBatch
      parameters:
      - name: planId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionBatchRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductionBatchDto'
...
  /api/v1/accounting/raw-materials:
    get:
      tags:
      - raw-material-controller
      operationId: listRawMaterials
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListRawMaterialDto'
    post:
      tags:
      - raw-material-controller
      operationId: createRawMaterial
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawMaterialRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseRawMaterialDto'
...
  /api/v1/hr/leave-requests/{id}/status:
    patch:
      tags:
      - hr-controller
      operationId: updateLeaveStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLeaveRequestDto'
  /api/v1/factory/production-plans/{id}/status:
    patch:
      tags:
      - factory-controller
      operationId: updatePlanStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductionPlanDto'
...
components:
  schemas:
...
    EmployeeRequest:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        hiredDate:
          type: string
          format: date
    ApiResponseEmployeeDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/EmployeeDto'
        timestamp:
          type: string
          format: date-time
    EmployeeDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        hiredDate:
          type: string
          format: date
    FactoryTaskRequest:
      required:
      - title
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignee:
          type: string
        status:
          type: string
        dueDate:
          type: string
          format: date
    ApiResponseFactoryTaskDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/FactoryTaskDto'
        timestamp:
          type: string
          format: date-time
    FactoryTaskDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        assignee:
          type: string
        status:
          type: string
        dueDate:
          type: string
          format: date
    ProductionPlanRequest:
      required:
      - planNumber
      - plannedDate
      - productName
      - quantity
      type: object
      properties:
        planNumber:
          type: string
        productName:
          type: string
        quantity:
          type: number
          format: double
        plannedDate:
          type: string
          format: date
        notes:
          type: string
    ApiResponseProductionPlanDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProductionPlanDto'
        timestamp:
          type: string
          format: date-time
    ProductionPlanDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        planNumber:
          type: string
        productName:
          type: string
        quantity:
          type: number
          format: double
        plannedDate:
          type: string
          format: date
        status:
          type: string
        notes:
          type: string
...
    RawMaterialRequest:
      required:
      - maxStock
      - minStock
      - name
      - reorderLevel
      - sku
      - unitType
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        unitType:
          type: string
        reorderLevel:
          type: number
        minStock:
          type: number
        maxStock:
          type: number
    ApiResponseRawMaterialDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/RawMaterialDto'
        timestamp:
          type: string
          format: date-time
    RawMaterialDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        sku:
          type: string
        unitType:
          type: string
        reorderLevel:
          type: number
        currentStock:
          type: number
        minStock:
          type: number
        maxStock:
          type: number
        stockStatus:
          type: string
...
    RawMaterialIntakeRequest:
      required:
      - batchCode
      - costPerUnit
      - quantity
      - rawMaterialId
      - unit
      type: object
      properties:
        rawMaterialId:
          type: integer
          format: int64
        batchCode:
          type: string
        quantity:
          type: number
        unit:
          type: string
        costPerUnit:
          type: number
        supplier:
          type: string
        notes:
          type: string
    ApiResponseRawMaterialBatchDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/RawMaterialBatchDto'
        timestamp:
          type: string
          format: date-time
    RawMaterialBatchDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        batchCode:
          type: string
        quantity:
          type: number
        unit:
          type: string
        costPerUnit:
          type: number
        supplier:
          type: string
        receivedAt:
          type: string
          format: date-time
        notes:
          type: string
    RawMaterialBatchRequest:
      required:
      - batchCode
      - costPerUnit
      - quantity
      - unit
      type: object
      properties:
        batchCode:
          type: string
        quantity:
          type: number
        unit:
          type: string
        costPerUnit:
          type: number
        supplier:
          type: string
        notes:
          type: string
    PayrollRunRequest:
      required:
      - initiatedBy
      - payrollDate
      type: object
      properties:
        payrollDate:
          type: string
          format: date
        initiatedBy:
          type: string
...
    ApiResponsePayrollRunDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/PayrollRunDto'
        timestamp:
          type: string
          format: date-time
    PayrollRunDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        runDate:
          type: string
          format: date
        status:
          type: string
        processedBy:
          type: string
        notes:
          type: string
    LeaveRequestRequest:
      required:
      - endDate
      - leaveType
      - startDate
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        leaveType:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
        status:
          type: string
    ApiResponseLeaveRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/LeaveRequestDto'
        timestamp:
          type: string
          format: date-time
    LeaveRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        leaveType:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductionBatchRequest:
      required:
      - batchNumber
      - quantityProduced
      type: object
      properties:
        batchNumber:
          type: string
        quantityProduced:
          type: number
          format: double
        loggedBy:
          type: string
        notes:
          type: string
    ApiResponseProductionBatchDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProductionBatchDto'
        timestamp:
          type: string
          format: date-time
    ProductionBatchDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        batchNumber:
          type: string
        quantityProduced:
          type: number
          format: double
        producedAt:
          type: string
          format: date-time
        loggedBy:
          type: string
        notes:
          type: string
...
    StatusRequest:
      type: object
      properties:
        status:
          type: string
...
    ApiResponseListRawMaterialBatchDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/RawMaterialBatchDto'
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListPayrollRunDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/PayrollRunDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListLeaveRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/LeaveRequestDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListEmployeeDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListFactoryTaskDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/FactoryTaskDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListProductionPlanDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductionPlanDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListProductionBatchDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductionBatchDto'
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListRawMaterialDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/RawMaterialDto'
        timestamp:
          type: string
          format: date-time
...